update Gatsby config to use absolute paths
