+++ gatsby-config.js
@@ -1,3 +1,5 @@
+const path = require('path');
+
 module.exports = {};
 
 module.exports = {
@@ -68,6 +70,18 @@ module.exports = {
       },
     },
     `gatsby-plugin-typescript`,
+    {
+      resolve: 'gatsby-plugin-root-import',
+      options: {
+        src: path.join(__dirname, 'src'),
+        assets: path.join(__dirname, 'src/assets'),
+        components: path.join(__dirname, 'src/components'),
+        domain: path.join(__dirname, 'src/domain'),
+        layout: path.join(__dirname, 'src/layout'),
+        pages: path.join(__dirname, 'src/pages'),
+        styles: path.join(__dirname, 'src/styles'),
+      },
+    },
     {
       resolve: `gatsby-plugin-react-redux`,
       options: {
