+++ package.json
@@ -103,6 +103,7 @@
     "chai": "^4.3.4",
     "cypress": "^7.7.0",
     "cypress-cucumber-preprocessor": "^4.1.1",
+    "gatsby-plugin-root-import": "^2.0.8",
     "gh-pages": "^3.2.0",
     "identity-obj-proxy": "^3.0.0",
     "ignore-styles": "^5.0.1",
